deployment_scripts/puppet/modules/dns_update/files/requirements.txt
requests>=2.0.0
dnspython>=1.12.0
kombu>=3.0.2
SQLAlchemy>=1.0.12
